+++ packages/apps/kadena-docs/src/hooks/useMediumScreen.ts
@@ -1,5 +1,5 @@
 import useMediaQuery from './useMediaQuery';
 
-export const useMediumScreen = () => {
-  return { hasMediumScreen: useMediaQuery('md') };
+export const useMediumScreen = (): boolean => {
+  return useMediaQuery('md');
 };
